Refactor compile benchmark profiling shortcut into a separate component
